requirements.txt
Django
django-rest-framework
Pillow
psycopg2-binary
python-decouple
drf-yasg